fix: validate product ID in deleteProduct controller function and handle errors with correct HTTP response code

+++ backend/controllers/product.controller.js
@@ -63,6 +63,12 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const productId = req.params.id;
 
+  if (!mongoose.isValidObjectId(productId)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(productId);
     res
@@ -70,6 +76,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("Error deleting product:", error);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Internal Server Error" });
   }
 };
